semantic-search-poc/requirements.txt
opensearch-py==2.4.2  # Official OpenSearch client
python-dotenv==1.0.0
rich==13.7.0  # For beautiful console output
typing-extensions==4.9.0
numpy>=1.24.0  # Vector math for similarity
inflect>=7.0.0  # Text pluralization/humanization used by generators
requests>=2.31.0  # For GraphQL API calls

# Embeddings (Cohere)
cohere>=5.0.0

# Bedrock support
boto3>=1.34.0

# LiteLLM (provider-unified OpenAI-compatible client)
litellm>=1.40.0
